+++ ThirdPartyAdapters/line/line/src/test/java/com/google/ads/mediation/line/LineInterstitialAdTest.kt
@@ -58,6 +58,7 @@ class LineInterstitialAdTest {
 
   @Test
   fun showAd_invokesFiveAdShowAd() {
+    lineInterstitialAd.loadAd()
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
 
     lineInterstitialAd.showAd(activity)
@@ -67,6 +68,8 @@ class LineInterstitialAdTest {
 
   @Test
   fun onFiveAdLoad_invokesOnSuccess() {
+    lineInterstitialAd.loadAd()
+
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
 
     verify(mockFiveAdInterstitial).setEventListener(lineInterstitialAd)
@@ -89,6 +92,7 @@ class LineInterstitialAdTest {
 
   @Test
   fun onClick_invokesReportAdClickedAndOnAdLeftApplication() {
+    lineInterstitialAd.loadAd()
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
 
     lineInterstitialAd.onClick(mockFiveAdInterstitial)
@@ -99,6 +103,7 @@ class LineInterstitialAdTest {
 
   @Test
   fun onFullScreenClose_invokesOnAdClosed() {
+    lineInterstitialAd.loadAd()
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
 
     lineInterstitialAd.onFullScreenClose(mockFiveAdInterstitial)
@@ -108,6 +113,7 @@ class LineInterstitialAdTest {
 
   @Test
   fun onImpression_invokesReportAdImpression() {
+    lineInterstitialAd.loadAd()
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
 
     lineInterstitialAd.onImpression(mockFiveAdInterstitial)
@@ -117,6 +123,7 @@ class LineInterstitialAdTest {
 
   @Test
   fun onViewError_invokesOnAdFailedToShow() {
+    lineInterstitialAd.loadAd()
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
     val dummyErrorCode = FiveAdErrorCode.INTERNAL_ERROR
     val adErrorCaptor = argumentCaptor<AdError>()
@@ -133,6 +140,7 @@ class LineInterstitialAdTest {
 
   @Test
   fun onFullScreenOpen_invokesOnAdOpened() {
+    lineInterstitialAd.loadAd()
     lineInterstitialAd.onFiveAdLoad(mockFiveAdInterstitial)
 
     lineInterstitialAd.onFullScreenOpen(mockFiveAdInterstitial)